massive refactor (see details)
1. Eradicated references to resources::manager (hereafter named **assets_manager**) from the logic code that is supposed to be deterministic and reliant exclusively on the state found within **cosmos** class.

**PROBLEM:**
Sometimes, part of the significant state is naturally influenced by state of the external assets like sound, particle effects, game image sizes etc.
For example, the lifetime (expressed in the number of logic steps) of an entity that stands for a sound effect (thus, containing components::sound_existence) would be naturally tied to the length of the sound file,
and likewise, the lifetime (expressed in the number of logic steps) of an entity that stands for a particle effect (thus, containing components::particles_existence) would be naturally tied to the maximum possible duration of the particle stream.
There is more. The prefabs:: and ingredients:: functions even rely on the sizes of the game images to create physical shapes.

The problem was that the logic code was getting these properties (sound lengths, sprite sizes, etc.) through a global resource manager obtained by a global call get_resources_manager (hereafter named "get_assets_manager"), a manager that was populated every launch with what was found in the files on the disk or the standard asset setters in code.
If one would like to save state of the **cosmos** along with some inputs for the purpose of replaying it later, it would mean that it is enough to replace a game image in the game's files to produce completely divergent results, rendering the savefile corrupted. This is unacceptable.

**SOLUTION:**
Each asset (a sound, a sprite, a particle effect, a shader, a spell) is required to expose a public member *get_logical_meta()* that returns the information that the logic code needs to know about a given asset. For example, a sprite would have a calculated physical shape (possibly from a polygonization file), a sound would have a calculated duration put into a single float etc.

All maps from asset ids into asset types are defined exactly once in the **assets_manager** via the **tuple_of_all_assets** typedef.
The template code will now automatically transform that tuple into a tuple of maps from asset identificators into a tuple of maps from asset identificators into **logical metadatas of assets** (more precisely, results of *get_logical_meta* call for each asset type), hereafter named **tuple_of_all_logical_metas_of_assets**.
If an asset type does not provide a *get_logical_meta* member function, it will simply not be added to that tuple (for example, logic does not require knowledge of shaders and physical textures (atlases)).

The **cosmos** is henceforth in possession of **tuple_of_all_logical_metas_of_assets**.

Furthermore, **assets_manager** provides a function (*write_logical_metas_of_assets_into*) that obtains current logical metas of all current assets and writes them once into the passed **cosmos**. If a resource changes during runtime (for example due to some GUI action in some future developer mode), it is enough to call that function once again to keep the **cosmos** in sync. Regenerating relevant state of entities due to changed logical metadatas is a concern for later, once it becomes possible to alter resources in runtime.

The **cosmos** will no longer need a global **assets_manager** to obtain the data it needs for logical processing, as it will have its own copy of logical metas.
Furthermore, the **cosmos** will be serialized (saved and loaded) along with these logical metas, meaning that when recording/replaying some saved **cosmos**, one does not need to worry about divergences due to some custom game sprites and sounds on the disk.
The only code that needs to call get_assets_manager() are the fragments dealing with GUI and viewing, and in general only the code that accepts cosmos by const&.

From now on, the server will not even need to have its own **assets_manager** to properly simulate and host the **cosmos**. It is enough that it has a proper savefile that contains the requisite logical metadatas.

It also plays nicely with the philosophies of separation of concerns, reduction of globals and whatnot.

2. Spell datas and physical materials became assets.

3. Delta encoders for trivial types (thus, encoding fixed-length sequences of bytes) will now automatically choose the index type (unsigned char, unsigned short, unsigned) depending of the size of the encoded type.

4. Moved all static_asserts regarding custom type traits into one file tests_of_traits.cpp

5. Massive improvements to the template code: replaced a template that takes a list of types and returns a **type** fulfilling a criterion with a template that takes a list of types and returns a **list of types** fulfilling a criterion, not just the first found type. find_matching_type was also implemented by means of filter_types. Several other helper templates have been added.

6. components::polygon is no longer in possession of uv values in the physical atlas space. It only holds uv values in local space (0.0-1.0) for the given game image id of whose it is also in possession. Previous design would effectively make polygons appear differently on the clients with differing positions of textures packed into atlas.

7. Fixed bugs in item_system.cpp related to pickup delay.

8. Eradicated renderable_positioning_type enum. Objects will be positioned by thier center universally.

9. Fixed introspection in for_each_child_entity_recursive when called for const entity handles.