Interp and performance fixes for low FPS clients.
client_time will fall behind current_time A LOT if fps is less than tickrate.
We must then snap the client_time so it is no further from current_time then delta.
    - otherwise even if the client speeds up suddenly, they'll always be behind
Also snap interpolation to 0-1 range because the current time might still get out of range if the logic step hasn't been run long enough.

Note that the client_time here is not the usual timer for how many steps should be performed.
We're using "if" instead of "while" when checking if it's less than current_time.
A nice side effect of that is that under less framerates, we'll re-predict the predicted world less in the future, relieving FPS a bit, showing a bit to the past.

The actual simulation steps performed are always determined by how many steps have arrived from the server.