Launch new thread when opening URLs
The open_url() function uses std::system() for opening a browser on the
given URL. Unfortunately, std::system() is a blocking call, rendering
the game unresponsive as long as the browser window is open. When
combined with the game window grabbing mouse input, this can make
it tricky for a user to close the browser and return to the game.